Moving BatchedVariable to utils
Before, we had issues with circular imports. By defining it in utils, we no
longer need to import things from base in utils, and from utils in base.
